// src/app/core/constants/enum.ts
export enum RoleAdminEnum {
  CompanyManager = 'CompanyManager',
}

export enum IRoleEnum {
  AgreementPrice = 1,
  Housing = 2,
  Procedure = 3,
  RentSalePrice = 4,
  SalePrice = 5,
  DelegationRequest = 6,
  ExternalWaiverRequest = 7,
  RecruitmentContract = 8,
  RentRequest = 9,
  ReturnRequest = 10,
  SponsorshipTransfer = 11,
  WaiverRequest = 12,
  ArrivalStation = 13,
  Branch = 14,
  City = 15,
  Company = 16,
  Country = 17,
  Detail = 18,
  ExperienceType = 19,
  Job = 20,
  Region = 21,
  Skill = 22,
  Template = 23,
  VisaType = 24,
  Customer = 25,
  Employee = 26,
  ExternalOffice = 27,
  ExternalOfficeUser = 28,
  WorkerBenefit = 29,
  Worker = 30,
  Setting = 31,
}

export enum IRolePermissionsEnum {
  canAdd = 'canAdd',
  canView = 'canView',
  canDelete = 'canDelete',
  canUpdate = 'canUpdate',
}